Add drawing circles mechanism.
Creates two separate functions for rendering a circle, one using
cartesian equation and the other using parametric equation. Which
function will be used to draw the circle is determined by one generic
function called "drawCircle".

Also, a debug mechanism is implemented. When activated, it shows in
the terminal messeges saying what is happening during the execution
of the program.