Adding vector to represent direction and speed.
Adds a vector that have a length proportional to the speed of the
circle and an angle compatible with the direction of the circle
movement. The 'a'/'d' keys can be used to increase/decrease the
angle. Also, the 'w'/'s' keys can be used to increase/decrease the
angle.