Changing the jump mechanism.
The jumping mechanism are reorganized in new functions. So, is
easier to implement the motion using more than one way. Until now,
just analytical integration is used to make the jump. This way use
either cartesian or parametric equation, depending on which is
currently switched on.

Also, the jumpingFlag is used to avoid changes in the speed or
angle during the jump.