Adding interactive viewing.
As the game is intended to be upgraded to a 3D game, the circle is
replaced by a sphere, created by the function glutWireSphere(). Besides
that, the camera follows this sphere with a third person perspective.
This perspective is correctly updated if the window is resized.

Mouse controls are included, making possible to zoom the camera by
right clicking and dragging up/down. Also, is possible to rotate
around Y axis by left clicking and dragging left/right as it is to
rotate around X axis by left clicking and dragging up/down.

Finally, a z frog coordinate is included, althought it doesn't make
any difference for the program now. In the future it will be used to
make possible to do 3D jumping.