Adding grid on the xz plane.
Renders a grid of size 100x100 and with 100x100 rows/columns.
Creates arrays of vertices to store this grid and other geometry
data. Furthermore, to prevent storing duplicate data, creates an
index array to save the order that the grid vertices need to be
rendered.

Also, fixes the problem that made camera rotations be around origin
instead of being around the sphere. However, still not possible to
correctly rotate around 2 different axes without one rotation
interfering on another. Because of that, one of the rotations is
a comment for now.