Rendering cube and reusing functions.
Creates a local initPrism() based on the old initCylinder() that
generalize the initialization of a prism. Then, is possible to render
both a cylinder and a cube calling the same function. Therefore, the
initCylinder() and the now created initCube() are made calling this.

Also, creates a local drawGeometry() that generalize the geometry draw.
This function is now called in all draw functions.