Creating a game mode and improving the game map.
Now the 'm' key switches between game mode and developer mode. On
the game mode the input are different, as follow:

- Mouse: Users don't need to click to move the camera. It can be
	rotated on the Y axis by moving the mouse right/left.
- <w, a, s, d> or <up arrow, left arrow, down arrow, rigth arrow>:
	Now the frog jump has velocity and fixe curvature angle.
	Then, with those keys users can jump to front, to left, to
	back or to right.

The cursor is invisible on the game mode.

In addition, changes the grid to make possible a future
implementation of road and river on the map. For this, part of the
grid is now lowered. Also, changes the side size of the grid's
triangles. On this way, the frog can jump exactly between the grid
rows.