Including a objLoader, new mesh system and render
Using mesh and render from the assignment example, for a more efficient
rendering. For now just the objects that comes from .obj file uses the
new mesh. and the other will be migrated later.

New folder res has been added to hold the resources: obj files and
textures.

As an example the frog is been drawn using a objfile.