Improving game mode and cars/logs mechanism.
Improves the game mode by setting all the core flags depending if
the context is on game mode or developer mode.

Also, improves the way that cars and logs are created, updated and
rendered. Now the cars and logs are rendered on the map more like
the traditional frogger game.