Fixing bugs.
Before this commit, sometimes the game entered on an infinit loop
while initializing. This was because it was triying to find a place
to render objects not so close to each other and that place didn't
exist. This bug is fixed by establishing a limit number of tries.

Also, if the frog was on a log that reached the end of screen, the
game was reseted but the user didn't lose points. This is fixed by
putting the mechanism that check this situation on the
detectCollision function.